requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-jazzmin==3.0.0
python-dotenv==1.0.0
Pillow==10.1.0
python-dateutil==2.8.2
gunicorn==21.2.0
